+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 ### Bug fixes
 
 * [#5241](https://github.com/bbatsov/rubocop/issues/5241): Fix an error for `Layout/AlignHash` when using a hash including only a keyword splat. ([@wata727][])
+* [#5247](https://github.com/bbatsov/rubocop/issues/5247): Fix undesirable
+  behavior in `Layout/CommentIndentation` when a comment is followed by an
+  outdented bare access modifier. ([@grantbdev][])
 
 ## 0.52.0 (2017-12-12)